Fix: ensure integer inputs update localization on language change
